+++ src/models/transaction_count_by_block_number.pb.gorm.go
@@ -21,7 +21,7 @@ var _ = fmt.Errorf
 var _ = math.Inf
 
 type TransactionCountByBlockNumberORM struct {
-	BlockNumber     uint32 `gorm:"index:transaction_count_by_block_count_idx_block_count"`
+	BlockNumber     uint64 `gorm:"index:transaction_count_by_block_count_idx_block_count"`
 	Count           uint32
 	TransactionHash string `gorm:"primary_key"`
 }
